+++ guillotina/tests/docker_containers/__init__.py
@@ -1,0 +1,3 @@
+from .cockroach import image as cockroach_image  # noqa
+from .etcd import image as etcd_image  # noqa
+from .postgres import image as postgres_image  # noqa