+++ guillotina/tests/docker_containers/etcd.py
@@ -1,26 +1,14 @@
-from time import sleep
+from guillotina.tests.docker_containers.base import BaseImage
 
 
-ETCD_IMAGE = 'quay.io/coreos/etcd:v3.2.0-rc.0'
-
-
-def run_docker_etcd(label='testingetcd'):
-    import docker
-    docker_client = docker.from_env(version='1.23')
-
-    # Clean up possible other docker containers
-    test_containers = docker_client.containers.list(
-        all=True,
-        filters={'label': label})
-    for test_container in test_containers:
-        test_container.stop()
-        test_container.remove(v=True, force=True)
-
-    # Create a new one
-    container = docker_client.containers.run(
+class ETCD(BaseImage):
+    label = 'etcd'
+    image = 'quay.io/coreos/etcd:v3.2.0-rc.0'
+    to_port = from_port = 2379
+    image_options = BaseImage.image_options.copy()
+    image_options.update(dict(
         name='my-etcd-1',
-        image=ETCD_IMAGE,
-        labels=[label],
+        mem_limit='200m',
         command=' '.join([
             '/usr/local/bin/etcd',
             '--name my-etcd-1',
@@ -33,40 +21,8 @@ def run_docker_etcd(label='testingetcd'):
             '--initial-cluster-token my-etcd-token',
             '--initial-cluster-state new',
             '--auto-compaction-retention 1'
-        ]),
-        detach=True,
-        ports={
-            '2379/tcp': 2379
-        },
-        cap_add=['IPC_LOCK'],
-        mem_limit='200m',
-        privileged=True
-    )
-    ident = container.id
-    count = 1
-
-    opened = False
-    host = ''
-
-    print('starting etcd')
-    while count < 30 and not opened:
-        count += 1
-        try:
-            docker_client.containers.get(ident)
-        except docker.errors.NotFound:
-            continue
-        sleep(1)
-    print('postgresql etcd')
-    return host
+        ])
+    ))
 
 
-def cleanup_etcd_docker(label='testingetcd'):
-    import docker
-    docker_client = docker.from_env(version='1.23')
-    # Clean up possible other docker containers
-    test_containers = docker_client.containers.list(
-        all=True,
-        filters={'label': label})
-    for test_container in test_containers:
-        test_container.kill()
-        test_container.remove(v=True, force=True)
+image = ETCD()